Enable song reordering and alert our internal datastructures of the new order.
This works fine if you haven't sorted on a column yet.  There seems to be no
way to put it back to "no ordering"

gtk_tree_model_sort_reset_default_sort_func doesn't seem to want ot work
with our song model, but once we figure out how to turn song sorting "off"
and let the user interact with it this will be complete and what's written
to the db will be consistent with what you see on screen.

We also need a way to alert our internal data structures when the column
sort buttons are pressed.  Any ideas?

Hushed two warnings in display.c also.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@62 f01d2545-417e-4e96-918e-98f8d0dbbcb6