Here are threaded reads/writes/md5ing/id3 modifying
Weirdnesses about it.  If you take the gtkpod_statusbar_message out of the
comments a weird A appears in your terminal and the app freezes.  I think
this has something to do with the main widget being set to not sensitive at
the time the status bar message request is made.  I tried fixing it with a
timer, but it's not really reliable.  Would it be ok to have return codes
for the functions handle_import() and handle_export()
I think this problem could be avoided that way.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@86 f01d2545-417e-4e96-918e-98f8d0dbbcb6