Feature request from Matt Adam <matt@phlegmatic.ca> to mount/unmount ipod on startup/shutdown.
Is there a way to get my module to autodetect my iPod(showing up as
/dev/sdX) just by plugging it in, and let me unplug it whenever I want(as
long as it's not mounted)  This new feature would be ultra cool if i could
do that.  Jorg?

There is a prefs option, etc etc.  we should document what functions need to
be modified in order to add new prefs as it's getting quite large.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@140 f01d2545-417e-4e96-918e-98f8d0dbbcb6