Gtkpod now displays the amount of free space your ipod has.
This also reverts the music directory code to use only FXX not fXX if
FXX does not exist.  Jorg will you look through to see if I revereted it
right?

I just tested all of this off of a fresh ipod and it seems to be working
wonderfully.

I'm gonna bump the version, and fix the readme to add the shortname
stuff to your /etc/fstab if you have problems.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@151 f01d2545-417e-4e96-918e-98f8d0dbbcb6