added date_parser2.l which was not easy because automake does not "really" support multiple lexers. I worked around it by
1) requiring "flex" -- it will not work with "lex"
2) defining
%option prefix= (...)
%option outfile="lex.yy.c"
3) hard-coding "LEX_OUTPUT_ROOT = lex.yy" in Makefile.am

date_parser2 will try to determine which kind of string we have, e.g.
<date1
date1 < d < date2
...

It should then break down the string into the date token and pass it
to the other date_parser. This is not done yet. (The original
date_parser is called with the entire string, so expect some error
messages).

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@200 f01d2545-417e-4e96-918e-98f8d0dbbcb6