* rewrite of id3 functions to use libid3tag rather then libid3. The result is much cleaner and simpler and removes the dependancy on libstdc++.
There might be a few bugs with charsets in here but I think I got
most of them.  Certainly the code is very simple and any bugs should
be easily fixable.

On debian libid3tag ships with a pkg-config file so this is how
I detect it in configure.in.  If other dists have trouble with this
we can fall back to traditional autoconf methods.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@272 f01d2545-417e-4e96-918e-98f8d0dbbcb6