IMPROVEMENT: Track number and total number of tracks is displayed as nn/nn in the track view, and can also be entered as such.
IMPROVEMENT: The tags for CD number and total number of CDs can
now be displayed (analoguous to track number and total number of
tracks above)

IMPROVEMENT: In the past, using programs other than gtkpod to
transfer files voided the extended information file. gtkpod will
not try to use the MD5 checksums stored in the extended
information file to match up the data when this occurs.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@304 f01d2545-417e-4e96-918e-98f8d0dbbcb6