Applied Olivier Crête's patch:
faad2 also provides a copy of libmp4v2. It works with gtkpod too, except
for a few includes that it requires but fails to pull and windows
compatibility macro. I've attached a patch to make gtkpod 0.72 compile
with faad2. I'll be using it in Gentoo (so we can allow both packages to
satisfy the dependency).

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@354 f01d2545-417e-4e96-918e-98f8d0dbbcb6