With support from  Armando Atienza started a way to register that tracks have been played externally (by xmms etc).
There will be a plugin (acutally the code is already here) that will
call "gtkpod -p <filename>" after a track has been played.

gtkpod will then contact a running instance of gtkpod and report that
filename via a UNIX socket (/tmp/gtkpod-<username>, comments welcome:
better use a pipe?). The running instance of gtkpod will increase the
playcount of that file by one. If the file is not in the database, it
is registered in ~/.gtkpod/offline_playcount.

If no running instance of gtkpod exists, the file is registered in
~/.gtkpod/offline_playcount.

TODO:

- After the import of the itunesDB the offline_playcount file must be
parsed.

- Write function get_track_by_ipod_filename() that finds a file in the
database if it is located on the iPod (as opposed to on the local
filesystem).

Most of the code is in clientserver.c.

Comments welcome.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@371 f01d2545-417e-4e96-918e-98f8d0dbbcb6