Changed behaviour of free space update as it was somewhat incompatible with the automounter: the free space update would mount the iPod every five seconds...
Now data from the disk is only read when you are working with gtkpod
(e.g. adding tracks or clicking on playlists). This avoids that the
iPod gets mounted while you are attempting to disconnect it. Also,
readint the disk data is done in a background job as accessing the
drive with the automounter could take more than 100 ms.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@402 f01d2545-417e-4e96-918e-98f8d0dbbcb6