IMPROVEMENT: Added new track information fields: iPod_Path, creation time, soundcheck, samplerate, "BPM"
IMPROVEMENT: Volume normalizing: when importing mp3 the replay
gain is read and the soundcheck field set accordingly. Use
the "soundcheck" feature of the iPod to have the volume
normalized during playback.
If no replay gain tag is set, use the "Normalize" as usual to
have mp3gain write a replay gain tag.

Added tooltips to describe the track colums (in prefs window
and main application window -- not the obvious ones like
"Artist", "Title" and the like).

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@411 f01d2545-417e-4e96-918e-98f8d0dbbcb6