Smart Playlists -- enjoy! But: I need testers!
Support for iTunes' "checked" field.

When sorting ignore 'the' and similar at the beginning of the
title (thanks to Chris Micacchi)

Use statvfs() instead of a call to the external command "df"
to determine the free space on the iPod (thanks to Steve Jay)

The fields "time_modified/played/created" can now be edited
from within the track view.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@458 f01d2545-417e-4e96-918e-98f8d0dbbcb6