BUGFIX: When changing the mountpoint, the change was not written to the preferences file correctly.
BUGFIX: 'Copy from iPod' just failed in offline mode instead of
giving meaningful error messages -> fixed.

IMPROVEMENT: 'Copy from iPod' is not 'Export from Database' and
allows to export from the local database as well. This is
useful if you want to rename files in a consistant manner.

Release as 0.93.1

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@599 f01d2545-417e-4e96-918e-98f8d0dbbcb6