2007-05-30  P.G. Richardson <phantom_sf at users.sourceforge.net>
* display_coverart.c

Changed the on_main_cover_image_clicked so that only the
tracks belonging to the album are displayed in the track
listing. The sorttabs are no longer updated. A lot
quicker and doesnt rely on the artist/album sorttabs
being visible.

* fetchcover.c

Status bar content changes in case album is not found.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@1511 f01d2545-417e-4e96-918e-98f8d0dbbcb6