2007-06-01  P.G. Richardson <phantom_sf at users.sourceforge.net>
* display_coverart.c

Added callback on display_coverart content panel for acting
on scroll events with the mouse wheel. Result is that moving
the scroll wheel while the mouse is over the coverart display
will scroll the covers in the respective direction.

Avoided possible null pointer in track_changed function.
Thanks to Daniele Forsi for raising it. May need further work
if this does not solve it.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@1516 f01d2545-417e-4e96-918e-98f8d0dbbcb6