* src/display_playlists.c (pm_drag_motion): don't allow drops into non-loaded repositories. (pm_create_treeview): don't use GTK_DEFAULT_MOTION as this gives optical feedback that drops into the column header is allowed, even though it is not.
* src/display_sorttabs.c (st_create_listview): renamed to
st_create_treeview(), simplified and merged with parts of
st_create_page which did part of creating the treeview. This
is the point when I realize that DND into the sort tabs is
not supported and I don't have to do anything. Code is
easier to read and understand now.

* src/display_tracks.c (tm_drag_motion): don't allow drops
into non-loaded repositories.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@1589 f01d2545-417e-4e96-918e-98f8d0dbbcb6