2007-06-27 P.G. Richardson <phantom_sf at users.sourceforge.net>
* src/details.c: tartwork_changed flag defined when a new
artwork file is selected from the filesystem.

* src/fetchcover.c: tartwork_changed flag added for all
occurrences of artwork being changed.

* src/display_coverart.c: reworking of the code to fix the
sizes of the covers according to the ipod thumbnail size.
When the pane is expanded, the coverart display's spacing
is widened rather than the individual covers being
stretched.
Instead of the highlights, reflections and covers images
being redrawn as black when null, the containing items
are hidden. They are redisplayed when a new cover image
is required.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@1621 f01d2545-417e-4e96-918e-98f8d0dbbcb6