2007-11-04 P.G. Richardson <phantom_sf at users.sourceforge.net>
* src/display_photos.c

New ipods do not necessarily have ITDB_THUMB_PHOTO_LARGE or any
other photo format so cannot assume these in display of thumbs in
iconview or preview components.

Fix is to find the first existing smalled thumb in the case of the
iconview and biggest in the case of preview and scale the pixbufs
to uniform sizes.

Also, fix to clear the preview when a data change like remove occurs.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@1757 f01d2545-417e-4e96-918e-98f8d0dbbcb6