2007-12-30 P.G. Richardson <phantom_sf at users.sourceforge.net>
* src/display_coverart.h
* src/display_coverart.c

Visible scaled images are cached to improve performance. Only when
they are no longer visible is the scaled version unreferenced. If the
scaled version does not exist then a new scaled version is created as
usual. Thus, when moving along the collection, only 1 image should
be scaled when the drawing occurs rather than 9.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@1888 f01d2545-417e-4e96-918e-98f8d0dbbcb6