Prefer Audiophile (Album) over Radio (Track) gain
Also adjust the ordering of the tags checked.  We now check the APE tag
first, because if a user has set it they likely would want to use that
value instead of the one added automatically by Lame at encode time.
Thanks to Tino Keitel for the patch.

This also resets the GainData struct in mp3_get_track_ape_replaygain(),
so that we don't end up using some mix of data from the LAME Tag and the
APE tag for ReplayGain.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@2323 f01d2545-417e-4e96-918e-98f8d0dbbcb6