Add preferences for handling ReplayGain
This adds two preference settings, replaygain_offset and
replaygain_mode_album_priority.  The first is a simple offset value that
is added to the ReplayGain value before being converted to a Soundcheck
value and stored in the iTunesDB.  The other setting allows a choice
between Audiofile (aka album) gain and Radio (aka track) gain.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@2325 f01d2545-417e-4e96-918e-98f8d0dbbcb6