libgpod expects empty strings to be NULL
gtkpod sets empty Itdb_Track strings to "" instead of NULL to make
things less error-prone. However, this should stay internal to gtkpod,
and as far as libgpod is concerned, empty fields should be NULL, not "".
This commit makes sure that empty fields will be set to NULL before
trying to write the iPod database.